выношу общую логику обновления
